requirements.txt
# Dashboard API
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Telegram Bot
python-telegram-bot==21.0

# Claude API
anthropic==0.40.0

# Gmail API
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Utils
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
python-dateutil==2.8.2
aiohttp==3.9.1
